--- a/cellrank/logging/__init__.py
+++ b/cellrank/logging/__init__.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+from cellrank.logging._logging import (
+    hint,
+    info,
+    debug,
+    warning,
+    print_versions,
+    print_version_and_date,
+)