--- a/cellrank/logging/__init__.py
+++ b/cellrank/logging/__init__.py
@@ -3,6 +3,7 @@ from cellrank.logging._logging import (
     hint,
     info,
     debug,
+    error,
     warning,
     print_versions,
     print_version_and_date,