--- a/cellrank/plotting/_cluster_lineages.py
+++ b/cellrank/plotting/_cluster_lineages.py
@@ -2,7 +2,7 @@
 """Cluster lineages module."""
 
 from types import MappingProxyType
-from typing import Dict, Tuple, Union, Optional, Sequence
+from typing import Dict, Tuple, Union, TypeVar, Optional, Sequence
 from pathlib import Path
 from collections import Iterable
 
@@ -10,15 +10,15 @@ import numpy as np
 
 import matplotlib.pyplot as plt
 
-from scanpy import logging as logg
-from anndata import AnnData
-
+from cellrank import logging as logg
 from cellrank.tools._utils import save_fig
 from cellrank.utils._utils import _get_n_cores, check_collection
 from cellrank.plotting._utils import _model_type, _create_models, _is_any_gam_mgcv
 from cellrank.tools._constants import LinKey
 from cellrank.utils._parallelize import parallelize
 from cellrank.utils.models._models import Model
+
+AnnData = TypeVar("AnnData")
 
 
 def _cl_process(
@@ -176,6 +176,7 @@ def cluster_lineage(
               containing the clustered genes.
     """
 
+    from anndata import AnnData as _AnnData
     import scanpy as sc
 
     lineage_key = str(LinKey.FORWARD if final else LinKey.BACKWARD)
@@ -214,7 +215,7 @@ def cluster_lineage(
         )(models, lineage, norm, **kwargs)
         logg.info("    Finish", time=start)
 
-        trends = AnnData(np.vstack(trends))
+        trends = _AnnData(np.vstack(trends))
         trends.obs_names = genes
 
         # sanity check