Implements #234 
Code:
```python
import cProfile
from pstats import SortKey, Stats

# This imports scanpy to current env, can't be run twice in a session
cProfile.run("import cellrank as cr", "import-cellrank.profile")
p = Stats("import-cellrank.profile")
p.sort_stats(SortKey.CUMULATIVE)
p.print_stats(30)
```
Results before:
```
Fri Jul  3 14:54:09 2020    import-cellrank.profile

         2224361 function calls (2189182 primitive calls) in 2.094 seconds

   Ordered by: cumulative time
   List reduced from 10688 to 30 due to restriction <30>

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
   3213/1    0.028    0.000    2.097    2.097 {built-in method builtins.exec}
   2589/1    0.012    0.000    2.097    2.097 <frozen importlib._bootstrap>:986(_find_and_load)
   2572/1    0.009    0.000    2.097    2.097 <frozen importlib._bootstrap>:956(_find_and_load_unlocked)
   2056/1    0.009    0.000    2.096    2.096 <frozen importlib._bootstrap>:650(_load_unlocked)
   1822/1    0.005    0.000    2.096    2.096 <frozen importlib._bootstrap_external>:777(exec_module)
   2786/1    0.002    0.000    2.095    2.095 <frozen importlib._bootstrap>:211(_call_with_frames_removed)
        1    0.000    0.000    2.095    2.095 /opt/projects/helmholtz/cellrank/cellrank/__init__.py:2(<module>)
        1    0.000    0.000    1.695    1.695 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/scanpy/__init__.py:2(<module>)
  830/144    0.002    0.000    1.402    0.010 {built-in method builtins.__import__}
2869/1163    0.004    0.000    1.251    0.001 <frozen importlib._bootstrap>:1017(_handle_fromlist)
        1    0.000    0.000    0.795    0.795 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/scanpy/_utils.py:1(<module>)
        1    0.000    0.000    0.689    0.689 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/anndata/__init__.py:1(<module>)
        1    0.000    0.000    0.666    0.666 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/scanpy/tools/__init__.py:1(<module>)
        1    0.000    0.000    0.486    0.486 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/scanpy/preprocessing/__init__.py:1(<module>)
        1    0.000    0.000    0.481    0.481 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/scanpy/preprocessing/_recipes.py:1(<module>)
        1    0.000    0.000    0.480    0.480 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/scanpy/preprocessing/_simple.py:1(<module>)
        1    0.000    0.000    0.480    0.480 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/anndata/_core/anndata.py:1(<module>)
        1    0.000    0.000    0.399    0.399 /opt/projects/helmholtz/cellrank/cellrank/tools/__init__.py:2(<module>)
     1822    0.014    0.000    0.302    0.000 <frozen importlib._bootstrap_external>:849(get_code)
        1    0.000    0.000    0.245    0.245 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/sklearn/__init__.py:1(<module>)
        1    0.000    0.000    0.243    0.243 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/sklearn/base.py:1(<module>)
        1    0.000    0.000    0.242    0.242 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/sklearn/utils/__init__.py:1(<module>)
        1    0.000    0.000    0.241    0.241 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/sklearn/utils/class_weight.py:5(<module>)
        1    0.000    0.000    0.240    0.240 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/sklearn/utils/validation.py:1(<module>)
        1    0.000    0.000    0.232    0.232 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/numba/__init__.py:1(<module>)
4307/4208    0.054    0.000    0.229    0.000 {built-in method builtins.__build_class__}
        1    0.000    0.000    0.224    0.224 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/sklearn/utils/fixes.py:1(<module>)
        1    0.000    0.000    0.223    0.223 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/scipy/stats/__init__.py:1(<module>)
        1    0.000    0.000    0.216    0.216 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/scipy/stats/stats.py:19(<module>)
        1    0.000    0.000    0.216    0.216 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/anndata/_core/raw.py:1(<module>)

```


Results after:
```
Fri Jul  3 14:53:15 2020    import-cellrank.profile

         2097111 function calls (2064590 primitive calls) in 1.824 seconds

   Ordered by: cumulative time
   List reduced from 9081 to 30 due to restriction <30>

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
   2206/1    0.025    0.000    1.827    1.827 {built-in method builtins.exec}
   2277/1    0.010    0.000    1.827    1.827 <frozen importlib._bootstrap>:986(_find_and_load)
   2260/1    0.008    0.000    1.827    1.827 <frozen importlib._bootstrap>:956(_find_and_load_unlocked)
   1758/1    0.007    0.000    1.826    1.826 <frozen importlib._bootstrap>:650(_load_unlocked)
   1530/1    0.004    0.000    1.826    1.826 <frozen importlib._bootstrap_external>:777(exec_module)
   2449/1    0.001    0.000    1.826    1.826 <frozen importlib._bootstrap>:211(_call_with_frames_removed)
        1    0.000    0.000    1.826    1.826 /opt/projects/helmholtz/cellrank/cellrank/__init__.py:2(<module>)
        1    0.000    0.000    1.612    1.612 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/scanpy/__init__.py:2(<module>)
  804/123    0.002    0.000    1.318    0.011 {built-in method builtins.__import__}
 2541/850    0.003    0.000    1.183    0.001 <frozen importlib._bootstrap>:1017(_handle_fromlist)
        1    0.000    0.000    0.766    0.766 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/scanpy/_utils.py:1(<module>)
        1    0.000    0.000    0.666    0.666 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/anndata/__init__.py:1(<module>)
        1    0.000    0.000    0.621    0.621 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/scanpy/tools/__init__.py:1(<module>)
        1    0.000    0.000    0.466    0.466 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/anndata/_core/anndata.py:1(<module>)
        1    0.000    0.000    0.452    0.452 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/scanpy/preprocessing/__init__.py:1(<module>)
        1    0.000    0.000    0.447    0.447 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/scanpy/preprocessing/_recipes.py:1(<module>)
        1    0.000    0.000    0.446    0.446 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/scanpy/preprocessing/_simple.py:1(<module>)
        1    0.000    0.000    0.235    0.235 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/sklearn/__init__.py:1(<module>)
        1    0.000    0.000    0.233    0.233 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/sklearn/base.py:1(<module>)
     1530    0.012    0.000    0.233    0.000 <frozen importlib._bootstrap_external>:849(get_code)
        1    0.000    0.000    0.233    0.233 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/sklearn/utils/__init__.py:1(<module>)
        1    0.000    0.000    0.232    0.232 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/sklearn/utils/class_weight.py:5(<module>)
        1    0.000    0.000    0.231    0.231 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/sklearn/utils/validation.py:1(<module>)
4111/4014    0.049    0.000    0.216    0.000 {built-in method builtins.__build_class__}
        1    0.000    0.000    0.215    0.215 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/anndata/_core/raw.py:1(<module>)
        1    0.000    0.000    0.215    0.215 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/sklearn/utils/fixes.py:1(<module>)
        1    0.000    0.000    0.214    0.214 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/scipy/stats/__init__.py:1(<module>)
        1    0.000    0.000    0.213    0.213 /opt/projects/helmholtz/cellrank/cellrank/tools/__init__.py:2(<module>)
        1    0.000    0.000    0.213    0.213 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/anndata/_core/aligned_mapping.py:1(<module>)
        1    0.000    0.000    0.211    0.211 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/anndata/_core/views.py:1(<module>)
```

This quick optimization saves me `0.3s`. @Marius1311 could you try it out?