@michalk8 2020-07-03
@Marius1311 I've added logging from scanpy (various reasons: more control, printing of versions, small speedup).
We could save ~`0.05` seconds not loading GPCCA immediately, but I've opted out for doing it early, because if the build fails, we know sooner. I've removed importing of `AnnData` and switched to TypeVar where possible.

With `scanpy` loaded, it takes about `0.250` seconds to import CellRank (scvelo: ~`0.2`), with `scvelo` loaded ~ `0.122`s.
We could further improving it by deferring it, but I advise against it (we use a lot of functions and it will be messy).

Here's the latest benchmark:
```
Fri Jul  3 16:26:24 2020    import-cellrank.profile

         2132924 function calls (2097840 primitive calls) in 1.732 seconds

   Ordered by: cumulative time
   List reduced from 9697 to 30 due to restriction <30>

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
   2355/1    0.025    0.000    1.735    1.735 {built-in method builtins.exec}
        1    0.000    0.000    1.735    1.735 <string>:1(<module>)
   2467/1    0.010    0.000    1.735    1.735 <frozen importlib._bootstrap>:986(_find_and_load)
   2449/1    0.008    0.000    1.735    1.735 <frozen importlib._bootstrap>:956(_find_and_load_unlocked)
   1915/1    0.007    0.000    1.735    1.735 <frozen importlib._bootstrap>:650(_load_unlocked)
   1656/1    0.004    0.000    1.734    1.734 <frozen importlib._bootstrap_external>:777(exec_module)
   2677/1    0.001    0.000    1.734    1.734 <frozen importlib._bootstrap>:211(_call_with_frames_removed)
        1    0.000    0.000    1.734    1.734 /opt/projects/helmholtz/cellrank/cellrank/__init__.py:2(<module>)
        1    0.000    0.000    1.481    1.481 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/scanpy/__init__.py:2(<module>)
  847/132    0.002    0.000    1.238    0.009 {built-in method builtins.__import__}
 2621/864    0.004    0.000    1.133    0.001 <frozen importlib._bootstrap>:1017(_handle_fromlist)
        1    0.000    0.000    0.725    0.725 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/scanpy/_utils.py:1(<module>)
        1    0.000    0.000    0.604    0.604 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/anndata/__init__.py:1(<module>)
        1    0.000    0.000    0.582    0.582 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/scanpy/tools/__init__.py:1(<module>)
        1    0.000    0.000    0.430    0.430 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/anndata/_core/anndata.py:1(<module>)
        1    0.000    0.000    0.421    0.421 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/scanpy/preprocessing/__init__.py:1(<module>)
        1    0.000    0.000    0.416    0.416 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/scanpy/preprocessing/_recipes.py:1(<module>)
        1    0.000    0.000    0.416    0.416 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/scanpy/preprocessing/_simple.py:1(<module>)
     1656    0.013    0.000    0.325    0.000 <frozen importlib._bootstrap_external>:849(get_code)
        1    0.000    0.000    0.252    0.252 /opt/projects/helmholtz/cellrank/cellrank/tools/__init__.py:2(<module>)
        1    0.000    0.000    0.218    0.218 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/sklearn/__init__.py:1(<module>)
        1    0.000    0.000    0.217    0.217 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/sklearn/base.py:1(<module>)
        1    0.000    0.000    0.216    0.216 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/sklearn/utils/__init__.py:1(<module>)
        1    0.000    0.000    0.214    0.214 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/sklearn/utils/class_weight.py:5(<module>)
        1    0.000    0.000    0.214    0.214 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/sklearn/utils/validation.py:1(<module>)
4574/4472    0.056    0.000    0.214    0.000 {built-in method builtins.__build_class__}
        1    0.000    0.000    0.200    0.200 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/sklearn/utils/fixes.py:1(<module>)
        1    0.000    0.000    0.199    0.199 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/scipy/stats/__init__.py:1(<module>)
        1    0.000    0.000    0.194    0.194 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/numba/__init__.py:1(<module>)
        1    0.000    0.000    0.193    0.193 /home/michal/.miniconda3/envs/cellrank/lib/python3.8/site-packages/scipy/stats/stats.py:19(<module>)
```