@Marius1311 2020-07-04
Thanks Mike! I will check this out today. Really nice to finally see codecov in action!!