@Marius1311 2020-07-04
Here are results on my machine, using a fresh miniconda environment with python 3.6. Note, I use `ipykernel` to create kernels that I can use to switch between different environments in my jupyter notebook, maybe that's slowing down things. 

# without these optimisations
- with scvelo and scanpy imported: 1s
- with only scanpy imported: 1.5s 
- without either imported: 4.5s

# with these optimisations
- with scvelo and scanpy imported: 0.6s
- with only scanpy imported: 0.9s 
- without either imported: 4.1s

This is perfectly sufficient! Happy to merge this.