@Marius1311 2020-07-04
logging also looks fine, cool to be able to print date and versions now!